Drastically improve test set results
By reimplementing the 'long' argument with only specified features,
unwanted long features are no longer trained.  For example, previously,
if a long RSI was useful but a long SMA was not, the user was forced to
train both long RSI and long SMA.  Now, the user can specify which long
features to train on.

In addition, the removal of the closing price from the training model
also improved results.  Closing prices are needed to evaluate the targets,
but they are not indicative of future trends.  Therefore, not training on
the closing price leads to higher accuracy in the model.